refactor: pkg/network migrated to internal/manager/p2p
the pkg is being used solely by manager, hence should have been an internal pkg

updated env config to assign a permanent peerID to manager, for easy identification in aggregator logs
