+++ .env.example
@@ -1,13 +1,9 @@
 # Alchemy or Infura RPC Endpoints, Public RPCs don't have enough rate limits
-# eth-holesky
-L1_RPC=
-# base-sepolia
-L2_RPC=
+L1_RPC=https://eth-holesky.g.alchemy.com/v2/E3OSaENxCMNoRBi_quYcmTNPGfRitxQa
+L2_RPC=https://base-sepolia.g.alchemy.com/v2/E3OSaENxCMNoRBi_quYcmTNPGfRitxQa
 
-# eth-holesky
-L1_WS_RPC=
-# base-sepolia
-L2_WS_RPC=
+L1_WS_RPC=wss://eth-holesky.g.alchemy.com/v2/E3OSaENxCMNoRBi_quYcmTNPGfRitxQa
+L2_WS_RPC=wss://base-sepolia.g.alchemy.com/v2/E3OSaENxCMNoRBi_quYcmTNPGfRitxQa
 
 L1_CHAIN=17000
 L2_CHAIN=84532
@@ -24,16 +20,19 @@ IPFS_HOST=aquamarine-urgent-limpet-846.mypinata.cloud
 # OTHENTIC Bootstrap Node
 OTHENTIC_BOOTSTRAP_ID=12D3KooWBNFG1QjuF3UKAKvqhdXcxh9iBmj88cM5eU2EK5Pa91KB
 OTHENTIC_BOOTSTRAP_SEED=97a64de0fb18532d4ce56fb35b730aedec993032b533f783b04c9175d465d9bf
+# LOG_DIR=data/logs/
 
 # PRODUCTION
 OTHENTIC_CLIENT_RPC_ADDRESS=http://localhost:8545
 MANAGER_IP_ADDRESS=http://localhost:9003
 DATABASE_IP_ADDRESS=http://localhost:9002
 AGGREGATOR_IP_ADDRESS=127.0.0.1
+HEALTH_IP_ADDRESS=http://localhost:9004
 
 AGGREGATOR_RPC_PORT=9001
 DATABASE_RPC_PORT=9002
 MANAGER_RPC_PORT=9003
+HEALTH_RPC_PORT=9004
 AGGREGATOR_P2P_PORT=9876
 DATABASE_DOCKER_PORT=9042
 DEBUG=true
@@ -44,5 +43,20 @@ ALCHEMY_API_KEY=
 PRIVATE_KEY_DEPLOYER=
 PRIVATE_KEY_AGGREGATOR=
 
+PRIVATE_KEY_MANAGER_P2P=
+
 # Currently not used, as Task Definitions are defined
-# PRIVATE_KEY_MANAGER=
+# PRIVATE_KEY_MANAGER=2f13b1ed304ba54824c27428b8a58e7c88bef75fe0b4387c90d69546af47f874
+
+
+# To run a Keeper
+# For Consensus Participation, Consensus Private Key from Docs
+PRIVATE_KEY=
+
+# For AVS participation, Controller Key from Docs
+OPERATOR_ADDRESS=
+OPERATOR_PRIVATE_KEY=
+
+# Ports available on the machine, default 9005 and 9006
+OPERATOR_RPC_PORT=
+OPERATOR_P2P_PORT=
